fix: restructure
