+++ src/app/seller-reviews/page.tsx
@@ -1,44 +1,14 @@
 "use client";
 
-import { Star, StarHalf, User } from "lucide-react";
+import { Search, Star, User } from "lucide-react";
 import Image from "next/image";
 import Link from "next/link";
-import { useSearchParams } from "next/navigation";
-import { useMemo, useState } from "react";
+import { useEffect, useState } from "react";
 
+import { Button } from "@/components/ui/Button";
 import { Card } from "@/components/ui/card/Card";
-
-// Mock data types
-interface Seller {
-  id: string;
-  name: string;
-  avatar?: string;
-  joinDate: string;
-  totalSales: number;
-  bio: string;
-}
-
-interface Review {
-  id: string;
-  sellerId: string;
-  rating: number;
-  comment: string;
-  createdAt: string;
-  buyer: {
-    id: string;
-    firstName: string;
-    lastName: string;
-    avatar?: string;
-  };
-  book: {
-    id: string;
-    title: string;
-    cover?: string;
-  };
-}
-
-// Mock data
-const mockSellers: Seller[] = [
+// Mock data for sellers
+const mockSellers = [
   {
     id: "seller1",
     name: "วรรณกรรม หนังสือดี",
@@ -46,6 +16,8 @@ const mockSellers: Seller[] = [
     joinDate: "2022-05-15",
     totalSales: 128,
     bio: "ร้านหนังสือที่รวบรวมวรรณกรรมคุณภาพ ทั้งไทยและต่างประเทศ เน้นสภาพหนังสือดี ราคาเป็นมิตร",
+    averageRating: 4.7,
+    totalReviews: 42,
   },
   {
     id: "seller2",
@@ -54,135 +26,90 @@ const mockSellers: Seller[] = [
     joinDate: "2021-10-08",
     totalSales: 256,
     bio: "จำหน่ายหนังสือมือสองสภาพดี ราคาถูก มีหนังสือหลากหลายประเภท ทั้งนิยาย การ์ตูน และตำราเรียน",
+    averageRating: 4.2,
+    totalReviews: 78,
+  },
+  {
+    id: "seller3",
+    name: "นิยายไทย",
+    avatar: "/placeholder.svg?height=80&width=80",
+    joinDate: "2022-01-20",
+    totalSales: 95,
+    bio: "ร้านหนังสือนิยายไทยคุณภาพ ทั้งนิยายรัก นิยายแฟนตาซี และนิยายสืบสวน",
+    averageRating: 4.9,
+    totalReviews: 36,
+  },
+  {
+    id: "seller4",
+    name: "ตำราเรียน มือหนึ่ง-มือสอง",
+    avatar: "/placeholder.svg?height=80&width=80",
+    joinDate: "2021-08-15",
+    totalSales: 312,
+    bio: "จำหน่ายตำราเรียนทุกระดับชั้น ทั้งมือหนึ่งและมือสอง ราคานักศึกษา",
+    averageRating: 4.5,
+    totalReviews: 124,
+  },
+  {
+    id: "seller5",
+    name: "หนังสือการ์ตูนญี่ปุ่น",
+    avatar: "/placeholder.svg?height=80&width=80",
+    joinDate: "2022-03-10",
+    totalSales: 187,
+    bio: "จำหน่ายการ์ตูนญี่ปุ่นทั้งภาษาไทยและภาษาญี่ปุ่น มีทั้งเล่มเดี่ยวและชุด",
+    averageRating: 4.6,
+    totalReviews: 67,
+  },
+  {
+    id: "seller6",
+    name: "หนังสือแนะแนวการศึกษา",
+    avatar: "/placeholder.svg?height=80&width=80",
+    joinDate: "2022-07-22",
+    totalSales: 76,
+    bio: "จำหน่ายหนังสือเตรียมสอบทุกระดับ ทั้ง O-NET, GAT/PAT, 9 วิชาสามัญ และอื่นๆ",
+    averageRating: 4.3,
+    totalReviews: 29,
+  },
+  {
+    id: "seller7",
+    name: "หนังสือภาษาอังกฤษ",
+    avatar: "/placeholder.svg?height=80&width=80",
+    joinDate: "2021-12-05",
+    totalSales: 143,
+    bio: "จำหน่ายหนังสือภาษาอังกฤษทั้งเพื่อการเรียนและความบันเทิง นำเข้าจากต่างประเทศ",
+    averageRating: 4.8,
+    totalReviews: 52,
+  },
+  {
+    id: "seller8",
+    name: "หนังสือพัฒนาตนเอง",
+    avatar: "/placeholder.svg?height=80&width=80",
+    joinDate: "2022-02-18",
+    totalSales: 209,
+    bio: "รวมหนังสือพัฒนาตนเอง ธุรกิจ การเงิน และจิตวิทยา ทั้งไทยและแปล",
+    averageRating: 4.4,
+    totalReviews: 83,
   },
 ];
 
-const mockReviews: Review[] = [
-  {
-    id: "review1",
-    sellerId: "seller1",
-    rating: 5,
-    comment: "หนังสือสภาพดีมาก ส่งเร็ว แพ็คดี คุ้มค่ากับราคา",
-    createdAt: "2023-11-15T08:30:00Z",
-    buyer: {
-      id: "buyer1",
-      firstName: "สมชาย",
-      lastName: "ใจดี",
-      avatar: "/placeholder.svg?height=40&width=40",
-    },
-    book: {
-      id: "book1",
-      title: "เจ้าชายน้อย",
-      cover: "/placeholder.svg?height=60&width=40",
-    },
-  },
-  {
-    id: "review2",
-    sellerId: "seller1",
-    rating: 4,
-    comment: "หนังสือสภาพดี แต่ส่งช้ากว่าที่คาดไว้นิดหน่อย",
-    createdAt: "2023-10-20T14:15:00Z",
-    buyer: {
-      id: "buyer2",
-      firstName: "สมหญิง",
-      lastName: "รักการอ่าน",
-      avatar: "/placeholder.svg?height=40&width=40",
-    },
-    book: {
-      id: "book2",
-      title: "แฮร์รี่ พอตเตอร์ กับศิลาอาถรรพ์",
-      cover: "/placeholder.svg?height=60&width=40",
-    },
-  },
-  {
-    id: "review3",
-    sellerId: "seller1",
-    rating: 5,
-    comment: "ประทับใจมาก ผู้ขายใส่ใจในรายละเอียด มีโน้ตเล็กๆ แถมมาด้วย",
-    createdAt: "2023-09-05T11:45:00Z",
-    buyer: {
-      id: "buyer3",
-      firstName: "วิชัย",
-      lastName: "ชอบอ่าน",
-      avatar: "/placeholder.svg?height=40&width=40",
-    },
-    book: {
-      id: "book3",
-      title: "คิดแบบเศรษฐศาสตร์",
-      cover: "/placeholder.svg?height=60&width=40",
-    },
-  },
-  {
-    id: "review4",
-    sellerId: "seller1",
-    rating: 3,
-    comment: "หนังสือโอเค แต่มีรอยพับที่มุมนิดหน่อย ไม่ได้ระบุในรายละเอียด",
-    createdAt: "2023-08-12T09:20:00Z",
-    buyer: {
-      id: "buyer4",
-      firstName: "นภา",
-      lastName: "ฟ้าใส",
-      avatar: "/placeholder.svg?height=40&width=40",
-    },
-    book: {
-      id: "book4",
-      title: "ประวัติศาสตร์ไทย",
-      cover: "/placeholder.svg?height=60&width=40",
-    },
-  },
-  {
-    id: "review5",
-    sellerId: "seller1",
-    rating: 5,
-    comment: "ดีมากครับ ส่งไว สภาพหนังสือเหมือนใหม่",
-    createdAt: "2023-07-28T16:50:00Z",
-    buyer: {
-      id: "buyer5",
-      firstName: "ธนา",
-      lastName: "ทรัพย์มาก",
-      avatar: "/placeholder.svg?height=40&width=40",
-    },
-    book: {
-      id: "book5",
-      title: "พ่อรวยสอนลูก",
-      cover: "/placeholder.svg?height=60&width=40",
-    },
-  },
-  {
-    id: "review6",
-    sellerId: "seller2",
-    rating: 4,
-    comment: "หนังสือสภาพดี ราคาเหมาะสม",
-    createdAt: "2023-11-10T10:30:00Z",
-    buyer: {
-      id: "buyer6",
-      firstName: "มานี",
-      lastName: "มีนา",
-      avatar: "/placeholder.svg?height=40&width=40",
-    },
-    book: {
-      id: "book6",
-      title: "มานี มีนา",
-      cover: "/placeholder.svg?height=60&width=40",
-    },
-  },
-];
-
-// Helper: Review stats
-const calculateReviewStats = (reviews: Review[]) => {
-  if (!reviews.length) return { average: 0, total: 0, counts: [] };
-  const total = reviews.length;
-  const sum = reviews.reduce((acc, review) => acc + review.rating, 0);
-  const average = sum / total;
-  const counts = [1, 2, 3, 4, 5].map((rating) => ({
-    rating,
-    count: reviews.filter((r) => r.rating === rating).length,
-    percentage: (reviews.filter((r) => r.rating === rating).length / total) * 100,
-  }));
-  return { average, total, counts };
+// Star Rating component
+const StarRating = ({ rating }: { rating: number }) => {
+  return (
+    <div className="flex items-center">
+      <div className="mr-1 flex">
+        {[...Array(5)].map((_, i) => (
+          <Star
+            key={i}
+            className={`h-4 w-4 ${i < Math.floor(rating) ? "fill-yellow-400 text-yellow-400" : "text-gray-300"}`}
+          />
+        ))}
+      </div>
+      <span className="text-sm font-medium">{rating.toFixed(1)}</span>
+    </div>
+  );
 };
 
-const formatDate = (dateString: string): string => {
+// Helper function to format date
+const formatDate = (dateString: string) => {
   const date = new Date(dateString);
   return date.toLocaleDateString("th-TH", {
     year: "numeric",
@@ -191,206 +118,150 @@ const formatDate = (dateString: string): string => {
   });
 };
 
-const StarRating: React.FC<{ rating: number }> = ({ rating }) => {
-  const fullStars = Math.floor(rating);
-  const hasHalfStar = rating % 1 >= 0.5;
-  return (
-    <div className="flex">
-      {[...Array(fullStars)].map((_, i) => (
-        <Star key={`full-${i}`} className="h-5 w-5 fill-yellow-400 text-yellow-400" />
-      ))}
-      {hasHalfStar && <StarHalf className="h-5 w-5 fill-yellow-400 text-yellow-400" />}
-      {[...Array(5 - fullStars - (hasHalfStar ? 1 : 0))].map((_, i) => (
-        <Star key={`empty-${i}`} className="h-5 w-5 text-gray-300" />
-      ))}
-    </div>
-  );
-};
+export default function SellerReviewsListPage() {
+  const [sellers, setSellers] = useState<typeof mockSellers>([]);
+  const [filteredSellers, setFilteredSellers] = useState<typeof mockSellers>([]);
+  const [searchQuery, setSearchQuery] = useState("");
+  const [loading, setLoading] = useState(true);
+  const [sortBy, setSortBy] = useState("rating"); // rating, reviews, sales, newest
 
-export default function SellerReviewsPage() {
-  const searchParams = useSearchParams();
-  const sellerId = searchParams.get("id") || "seller1";
+  useEffect(() => {
+    // Simulate API call to fetch sellers
+    setLoading(true);
 
-  const [sortBy, setSortBy] = useState<string>("newest");
-
-  const seller = useMemo(() => mockSellers.find((s) => s.id === sellerId) || null, [sellerId]);
-
-  const sellerReviews = useMemo(() => mockReviews.filter((r) => r.sellerId === sellerId), [sellerId]);
-
-  const reviews = useMemo(() => {
-    return [...sellerReviews].sort((a, b) => {
-      switch (sortBy) {
-        case "newest":
-          return new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime();
-        case "oldest":
-          return new Date(a.createdAt).getTime() - new Date(b.createdAt).getTime();
-        case "highest":
-          return b.rating - a.rating;
-        case "lowest":
-          return a.rating - b.rating;
-        default:
-          return 0;
-      }
+    // Sort sellers based on sortBy
+    const sortedSellers = [...mockSellers].sort((a, b) => {
+      if (sortBy === "rating") return b.averageRating - a.averageRating;
+      if (sortBy === "reviews") return b.totalReviews - a.totalReviews;
+      if (sortBy === "sales") return b.totalSales - a.totalSales;
+      if (sortBy === "newest") return new Date(b.joinDate).getTime() - new Date(a.joinDate).getTime();
+      return 0;
     });
-  }, [sellerReviews, sortBy]);
 
-  const stats = useMemo(() => calculateReviewStats(sellerReviews), [sellerReviews]);
+    setSellers(sortedSellers);
+    setFilteredSellers(sortedSellers);
+    setLoading(false);
+  }, [sortBy]);
 
-  if (!seller) {
+  useEffect(() => {
+    // Filter sellers based on search query
+    if (searchQuery.trim() === "") {
+      setFilteredSellers(sellers);
+    } else {
+      const filtered = sellers.filter(
+        (seller) =>
+          seller.name.toLowerCase().includes(searchQuery.toLowerCase()) ||
+          seller.bio.toLowerCase().includes(searchQuery.toLowerCase())
+      );
+      setFilteredSellers(filtered);
+    }
+  }, [searchQuery, sellers]);
+
+  if (loading) {
     return (
-      <div className="container mx-auto px-4 py-8 text-center">
-        <h1 className="text-2xl font-bold text-red-500">ไม่พบข้อมูลผู้ขาย</h1>
-        <p className="mt-4">
-          <Link href="/search" className="text-blue-500 hover:underline">
-            กลับไปยังหน้าค้นหา
-          </Link>
-        </p>
+      <div className="container mx-auto px-4 py-8">
+        <div className="flex h-64 items-center justify-center">
+          <div className="h-12 w-12 animate-spin rounded-full border-b-2 border-t-2 border-blue-500"></div>
+        </div>
       </div>
     );
   }
 
-  // ... UI rendering from original code (not repeated here for brevity)
   return (
     <div className="container mx-auto px-4 py-8">
-      <h1 className="mb-6 text-3xl font-bold">รีวิวผู้ขาย</h1>
-      {/* Seller Profile */}
-      <Card className="mb-8">
-        <div className="flex flex-col items-start gap-6 p-4 md:flex-row md:items-center">
-          <div className="relative h-20 w-20 overflow-hidden rounded-full bg-gray-100">
-            {seller.avatar ? (
-              <Image src={seller.avatar || "/placeholder.svg"} alt={seller.name} fill className="object-cover" />
-            ) : (
-              <div className="flex h-full w-full items-center justify-center">
-                <User className="h-12 w-12 text-gray-400" />
-              </div>
-            )}
-          </div>
+      <h1 className="mb-6 text-3xl font-bold">ผู้ขายทั้งหมด</h1>
 
-          <div className="flex-1">
-            <h1 className="text-2xl font-bold">{seller.name}</h1>
-            <p className="mt-1 text-gray-600">สมาชิกตั้งแต่ {formatDate(seller.joinDate)}</p>
-            <p className="text-gray-600">ขายแล้ว {seller.totalSales} รายการ</p>
-            <p className="mt-2">{seller.bio}</p>
+      {/* Search and Filter Controls */}
+      <div className="mb-8 flex flex-col gap-4 md:flex-row">
+        <div className="relative flex-1">
+          <div className="pointer-events-none absolute inset-y-0 left-0 flex items-center pl-3">
+            <Search className="h-5 w-5 text-gray-400" />
           </div>
-
-          <div className="flex flex-col items-center rounded-lg bg-gray-50 p-4">
-            <div className="text-3xl font-bold text-blue-600">{stats.average.toFixed(1)}</div>
-            <StarRating rating={stats.average} />
-            <div className="mt-1 text-sm text-gray-500">{stats.total} รีวิว</div>
-          </div>
+          <input
+            type="text"
+            placeholder="ค้นหาร้านค้า..."
+            value={searchQuery}
+            onChange={(e) => setSearchQuery(e.target.value)}
+            className="w-full rounded-md border py-2 pl-10 pr-4 focus:outline-none focus:ring-2 focus:ring-blue-500"
+          />
         </div>
-      </Card>
 
-      {/* Rating Statistics */}
-      <Card className="mb-8">
-        <div className="p-4">
-          <h2 className="mb-4 text-xl font-semibold">สถิติคะแนน</h2>
-
-          <div className="space-y-2">
-            {stats.counts
-              .slice()
-              .reverse()
-              .map((stat) => (
-                <div key={stat.rating} className="flex items-center gap-2">
-                  <div className="w-16 text-sm font-medium">{stat.rating} ดาว</div>
-                  <div className="h-4 flex-1 overflow-hidden rounded-full bg-gray-200">
-                    <div className="h-full rounded-full bg-yellow-400" style={{ width: `${stat.percentage}%` }}></div>
-                  </div>
-                  <div className="w-16 text-right text-sm">{stat.count} รีวิว</div>
-                </div>
-              ))}
-          </div>
+        <div className="w-full md:w-48">
+          <select
+            value={sortBy}
+            onChange={(e) => setSortBy(e.target.value)}
+            className="w-full rounded-md border px-3 py-2 focus:outline-none focus:ring-2 focus:ring-blue-500"
+          >
+            <option value="rating">เรียงตามคะแนน</option>
+            <option value="reviews">เรียงตามจำนวนรีวิว</option>
+            <option value="sales">เรียงตามยอดขาย</option>
+            <option value="newest">เรียงตามวันที่เข้าร่วม</option>
+          </select>
         </div>
-      </Card>
-
-      {/* Sort Controls */}
-      <div className="mb-4 flex justify-end">
-        <select
-          value={sortBy}
-          onChange={(e) => setSortBy(e.target.value)}
-          className="rounded-md border px-3 py-2 focus:outline-none focus:ring-2 focus:ring-blue-500"
-        >
-          <option value="newest">ล่าสุด</option>
-          <option value="oldest">เก่าสุด</option>
-          <option value="highest">คะแนนสูงสุด</option>
-          <option value="lowest">คะแนนต่ำสุด</option>
-        </select>
       </div>
 
-      {/* Reviews List */}
-      {reviews.length === 0 ? (
-        <Card>
-          <div className="p-8 text-center">
-            <h3 className="text-xl font-medium text-gray-500">ยังไม่มีรีวิว</h3>
-            <p className="mt-2 text-gray-400">เมื่อมีลูกค้ารีวิว รีวิวจะแสดงที่นี่</p>
-          </div>
-        </Card>
+      {/* Sellers Grid */}
+      {filteredSellers.length === 0 ? (
+        <div className="py-12 text-center">
+          <h3 className="text-xl font-medium text-gray-500">ไม่พบร้านค้าที่ตรงกับการค้นหา</h3>
+          <p className="mt-2 text-gray-400">ลองค้นหาด้วยคำอื่น หรือล้างการค้นหา</p>
+          <Button
+            onClick={() => {
+              setSearchQuery("");
+            }}
+            className="mt-4 rounded-md bg-blue-500 px-4 py-2 text-white transition hover:bg-blue-600"
+          >
+            ล้างการค้นหา
+          </Button>
+        </div>
       ) : (
-        <div className="space-y-4">
-          {reviews.map((review) => (
-            <Card key={review.id} className="overflow-hidden">
-              <div className="p-4">
-                {/* Review Header */}
-                <div className="mb-4 flex items-start justify-between">
-                  <div className="flex items-center gap-3">
-                    <div className="relative h-10 w-10 overflow-hidden rounded-full bg-gray-100">
-                      {review.buyer.avatar ? (
-                        <Image
-                          src={review.buyer.avatar || "/placeholder.svg"}
-                          alt={`${review.buyer.firstName} ${review.buyer.lastName}`}
-                          fill
-                          className="object-cover"
-                        />
-                      ) : (
-                        <div className="flex h-full w-full items-center justify-center">
-                          <User className="h-6 w-6 text-gray-400" />
-                        </div>
-                      )}
-                    </div>
-                    <div>
-                      <div className="font-medium">
-                        {review.buyer.firstName} {review.buyer.lastName}
+        <div className="grid grid-cols-1 gap-6 md:grid-cols-2 lg:grid-cols-3">
+          {filteredSellers.map((seller) => (
+            <Card key={seller.id} className="overflow-hidden transition hover:shadow-lg">
+              <div className="p-5">
+                <div className="flex items-start gap-4">
+                  <div className="relative h-16 w-16 flex-shrink-0 overflow-hidden rounded-full bg-gray-100">
+                    {seller.avatar ? (
+                      <Image
+                        src={seller.avatar || "/placeholder.svg"}
+                        alt={seller.name}
+                        fill
+                        className="object-cover"
+                      />
+                    ) : (
+                      <div className="flex h-full w-full items-center justify-center">
+                        <User className="h-8 w-8 text-gray-400" />
                       </div>
-                      <div className="text-sm text-gray-500">{formatDate(review.createdAt)}</div>
+                    )}
+                  </div>
+
+                  <div className="flex-1">
+                    <h2 className="text-xl font-semibold">{seller.name}</h2>
+                    <div className="mt-1 flex items-center">
+                      <StarRating rating={seller.averageRating} />
+                      <span className="ml-2 text-sm text-gray-500">({seller.totalReviews} รีวิว)</span>
                     </div>
+                    <p className="mt-1 text-sm text-gray-500">สมาชิกตั้งแต่ {formatDate(seller.joinDate)}</p>
                   </div>
-                  <StarRating rating={review.rating} />
                 </div>
 
-                {/* Book Info */}
-                <div className="mb-4 flex items-center gap-3 rounded-md bg-gray-50 p-3">
-                  <div className="h-15 relative w-10 overflow-hidden">
-                    <Image
-                      src={review.book.cover || "/placeholder.svg"}
-                      alt={review.book.title}
-                      width={40}
-                      height={60}
-                      className="object-cover"
-                    />
-                  </div>
-                  <div className="font-medium">{review.book.title}</div>
-                </div>
+                <div className="mt-4 line-clamp-2 text-sm text-gray-600">{seller.bio}</div>
 
-                {/* Review Comment */}
-                <div className="text-gray-700">{review.comment}</div>
+                <div className="mt-4 flex items-center justify-between">
+                  <div className="text-sm">
+                    <span className="font-medium">{seller.totalSales}</span> รายการขายแล้ว
+                  </div>
+
+                  {/*TODO: Implement dynamic link seller-reviews page */}
+                  <Link href="/seller-reviews" passHref legacyBehavior>
+                    <Button asChild className="text-sm font-medium">
+                      <a>ดูรีวิวทั้งหมด</a>
+                    </Button>
+                  </Link>
+                </div>
               </div>
             </Card>
           ))}
-        </div>
-      )}
-
-      {/* Pagination - would be implemented with real data */}
-      {reviews.length > 0 && (
-        <div className="mt-8 flex justify-center">
-          <nav className="flex items-center gap-1">
-            <button className="rounded border px-3 py-1 hover:bg-gray-100 disabled:opacity-50" disabled>
-              &laquo; ก่อนหน้า
-            </button>
-            <button className="rounded border bg-blue-500 px-3 py-1 text-white">1</button>
-            <button className="rounded border px-3 py-1 hover:bg-gray-100">2</button>
-            <button className="rounded border px-3 py-1 hover:bg-gray-100">3</button>
-            <button className="rounded border px-3 py-1 hover:bg-gray-100">ถัดไป &raquo;</button>
-          </nav>
         </div>
       )}
     </div>
